[refactor] StarImageDialog component 패키지로 이동
