+++ feature/home/src/main/kotlin/com/unifest/android/feature/home/HomeScreen.kt
@@ -48,6 +48,7 @@ import com.unifest.android.feature.festival.viewmodel.FestivalViewModel
 import com.unifest.android.feature.home.component.Calendar
 import com.unifest.android.feature.home.component.FestivalScheduleItem
 import com.unifest.android.feature.home.component.IncomingFestivalCard
+import com.unifest.android.feature.home.component.StarImageDialog
 import com.unifest.android.feature.home.preview.HomePreviewParameterProvider
 import com.unifest.android.feature.home.viewmodel.HomeUiAction
 import com.unifest.android.feature.home.viewmodel.HomeUiEvent
